requirements.txt
fastapi
uvicorn[standard]
pydantic

# Exporter for Google Cloud
opentelemetry-exporter-gcp-trace

# ML Libraries for Iris Model
scikit-learn
pandas
numpy